+++ app/build.gradle
@@ -48,6 +48,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
     implementation "com.google.code.gson:gson:2.10"
     implementation 'androidx.room:room-common:2.4.2'
+    implementation 'androidx.room:room-ktx:2.5.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
