+++ app/src/main/java/com/ikalne/meetmap/bbddfavlist/EventDao.kt
@@ -7,16 +7,16 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 
 @Dao
-interface EventDao{
+interface EventDao {
     @Query("SELECT * FROM event_table ORDER BY ev_date ASC")
     fun getAlphabetizedEvent(): LiveData<List<Event>>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(Event: Event)
+    fun insert(event: Event)
 
     @Query("DELETE FROM event_table")
     fun deleteAll()
 
     @Query("DELETE FROM event_table WHERE ev_id = :ev_id")
-    fun delete(ev_id: String)
+    fun delete(ev_id: Int)
 }
