+++ app/src/main/java/com/ikalne/meetmap/bbddfavlist/EventViewHolder.kt
@@ -1,6 +1,8 @@
 package com.ikalne.meetmap.bbddfavlist
 
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.ikalne.meetmap.R
@@ -19,7 +21,9 @@ class EventViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     }
 
     companion object {
-        fun create(view: View): EventViewHolder {
+        fun create(parent: ViewGroup): EventViewHolder {
+            val view: View = LayoutInflater.from(parent.context)
+                .inflate(R.layout.favlist_item, parent, false)
             return EventViewHolder(view)
         }
     }
